CMake based build system added
The code was partly taken from hermes3d, partly written by myself. This patch
only adds the CMake files, but doesn't remove the old system (this will be done
in the next patch).

The new CMake based build system has the same functionality as before, e.g. you
can build real or complex (or both) versions of the library, with or without
debugging symbols, with or without examples, etc.

Signed-off-by: David Andrs <andrsd@gmail.com>