Call the real hermes2d library as libhermes2d-real
This is needed, so that it doesn't conflict with the python target for the
hermes2d python module. The Python module is called hermes2d.so and the CMake
target is just hermes2d (and the prefix is set to "", see cmake/UseCython) so
it conflicts if the C++ library is called libhermes2d.so

Also it is now more consistent, as the complex library is named
"libhermes2d-cplx".