forms.pyx: the "void *data" parameter simulated in Cython
That's because this is not yet implemented in the C++ hermes2d. So we just
hardwire several special cases, that are needed in tests.

The real fix should be to implmenet this in hermes C++, but that will require
some substantial changes to the assembling procedure, so this should be done
later, when all the Python stuff works fine.

Also the example 08 was fixed.