Python updated to build with recent hermes
Originally I had about 50+ patches, but it was almost impossible to rebase them
to the newest hermes, so I just took the python directory as a whole and copied
it.

The most important changes are:

* CMake build system implemented for the python wrappers
* DiscreteProblem was converted to WeakForm
* all examples were fixed
* the tests were improved
* all cython files moved to python/hermes2d
* the ScalarView can now plot using mayavi or matplotlib
* the finalize() function removed

And lots of small fixes. There are still some little glitches, that are fixed
in the following patches (they deserve to have their own patch to easy future
debugging, if the fix was not the right one).