WITH_GLUT option to CMake implemented
If WITH_GLUT=no is set, then hermes2d is compiled with the "-DNOGLUT"
preprocessor directive.

A few changes had to be made in scr/view.h and src/hermes2d.cpp, mainly just
commenting out calls to symbols like glut_init() and finish_glut_main_loop(),
that should not be called if NOGLUT is set, otherwise the linker complains.

Signed-off-by: Ondrej Certik <ondrej@certik.cz>