Python wrappers updated to new forms
Now everything works and tests pass.

The example 05 was updated (fixed), previously it used surf_int_G_v in the
boundary condition and it was wrong. The C++ version was fixed some time ago,
so now the python version has to be fixed as well. This change made the
solution a bit different, so the corresponding test was updated as well.

Problems introduced by going to new forms with regards to Python:

1) The tests now run more than 2x slower (from 3.89 to 9.70)
2) All integrals are integrated with the order 20 (this probably causes the
slowdown in 1). That's because there is no easy way to implement all the forms
in Cython and use templates (e.g. it would have to be implemented in C++ for
any example that the user wants to do).

Signed-off-by: Ondrej Certik <ondrej@certik.cz>
Signed-off-by: Pavel Solin <solin@unr.edu>