Mesh load/save capability moved into H2DReader class
Loading and saving the mesh should be done via MeshLoader classes. There
is an ExodusII mesh loader class already. This patch changes the load/save
functionality the same way as it is done for ExodusII (i.e. one has to instantiate
the mesh loader class and then load the mesh). That gives more flexibility
on what formats can be pulled in in the future.

The old interface is preserved and marked deprecated. Anyone who will use it
gets a warning that they should migrate to the new API and that this old
API will be removed in the future.