Tutorial example 15 improved.
I added a parameter NEWTON_ON_COARSE MESH that allows the user to
compare the difference between solving the coarse mesh problem
after each adaptivity step or just using a projection of the last
fine mesh solution. Of course the latter is faster and it turns out
that the DOF convergence is not affected.