Adapt::Adapt() now takes pointer to LinSystem / NonlinSystem.
Formerly it took pointers to all coarse mesh spaces. This change
makes it possible to handle the spaces in a more compact way and
to use in the Adapt class more functionality of LinSystem, such
as the weak forms or the solver. Thus it will be possible, at least
in the simplest cases, to have the Adapt class do everything for
the user, including the computation of the reference solution,
calculation of errors, and refinement.