Unification of bilin/lin and jac/res forms now complete.
Hermes only has two types of forms: matrix forms and vector forms.
Also a bug fixed: For RefSystem, nonlinear solve was always called,
this was due to some mismathed use of virtual methods. I introduced
a new variable "linear" to LinSystem which is set to true in the
constructor of LinSystem and to false in the constructor of
NonlinSystem. In order to ascertain whether RefSystem is linear
or nonlinear, I am asking the value of this variable in this->base,
rather than calling any virtual methods. This works reliably.