Assembling in FeProblem split into two layers.
The basic assemble() takes a Tuple of Solutions (before it
expected the coefficient vector and converted the vector
into the solutions inside). Now there is a thin layer
above the basic assemble() that takes the coefficient
vector, converts it to a Tuple of solutions, and passes
the Tuple into the basic assemble() function. In other
words, now assemble() can be used with both the coefficient
vector or with a Tuple of solutions.