Unification of LinSystem and NonlinSystem - step 1.
New class DiscreteProblem introduced which is nonlinear by default.
Its special case LinearProblem was created to handle linear cases.
The default assemble() function in DiscreteProblem assembles and
exposes to the user the matrix, the vector Dir, and the vecton RHS.
This function used for nonlinear problems as-is, and for linear
problems the Dir vector is added to the right-hand side.