Automatic vertical scaling in ScalarView updated.
It now performs automatic scaling not only if the function has a too big range,
but also if the range is too small. In the first case, it is scaled down so that
it does not appear as a thin tall object. In the other case it is scaled
up so that considerable amount of detail is visible.