Function solve_newton_adapt() created in example 15.
Other changes made, including cancelling classes H1Adapt, L2Adapt
and HcurlAdapt. These differed is the projection forms only, so
now these are set in the constructor of Adapt. Important consequence
of this is that we can do adaptivity when different solution
components lie in different spaces (L2, H1, Hcurl, Hdiv). This
was not possible before.